+++ HealthApp/package.json
@@ -13,6 +13,7 @@
     "@react-navigation/native-stack": "^6.6.1",
     "expo": "~44.0.2",
     "expo-splash-screen": "~0.14.1",
+    "expo-sqlite": "~10.1.0",
     "expo-status-bar": "~1.2.0",
     "react": "17.0.1",
     "react-dom": "17.0.1",
