+++ HealthApp/src/components/mainButton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button, Pressable, StyleSheet, Text, View } from 'react-native';
+import { Pressable, StyleSheet, Text, View } from 'react-native';
 
 const IndexButton = (props) => {
     return (
