+++ HealthApp/src/screens/HomeScreen.js
@@ -2,17 +2,26 @@ import { StatusBar } from 'expo-status-bar';
 import React from 'react';
 import { View, Text } from 'react-native';
 import { StyleSheet } from 'react-native';
-import IndexButton from './components/mainButton';
+import IndexButton from '../components/mainButton';
 
 export default function HomeScreen({navigation}) {
     const moveToExercise = () => {
         navigation.navigate('ExerciseScreen')
     }
+
+    const moveToWater = () => {
+        navigation.navigate('WaterScreen')
+    }
+
+    const moveToCalories = () => {
+        navigation.navigate('CalorieScreen')
+    }
+
     return (
         <View style={styles.container}>
             <IndexButton label={'Exercise'} nav = {moveToExercise}></IndexButton>
-            <IndexButton label={'Water'}></IndexButton>
-            <IndexButton label={'Calories'}></IndexButton>
+            <IndexButton label={'Water'} nav = {moveToWater}></IndexButton>
+            <IndexButton label={'Calories'} nav = {moveToCalories}></IndexButton>
             <StatusBar style="auto" />
         </View>
     );
