+++ HealthApp/src/screens/WaterScreen.js
@@ -1,0 +1,13 @@
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { View, Text } from 'react-native';
+import { StyleSheet } from 'react-native';
+
+export default function WaterScreen({navigation}) {
+
+    return (
+        <View>
+            <Text>PlaceHolder</Text>
+        </View>
+    );
+}
